refactored new palette form navbar to its own component
