+++ src/components/PaletteFormNav/index.js
@@ -1,0 +1,96 @@
+import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import { styled } from "@mui/material/styles";
+import { Button } from "@mui/material";
+import CssBaseline from "@mui/material/CssBaseline";
+import MuiAppBar from "@mui/material/AppBar";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import IconButton from "@mui/material/IconButton";
+import MenuIcon from "@mui/icons-material/Menu";
+import { ValidatorForm, TextValidator } from "react-material-ui-form-validator";
+
+const drawerWidth = 400;
+
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+  transition: theme.transitions.create(["margin", "width"], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    width: `calc(100% - ${drawerWidth}px)`,
+    marginLeft: `${drawerWidth}px`,
+    transition: theme.transitions.create(["margin", "width"], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
+
+class PaletteFormNav extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { newPaletteName: "" };
+    this.handleChange = this.handleChange.bind(this);
+  }
+  componentDidMount() {
+    ValidatorForm.addValidationRule("isPaletteNameUnique", (value) =>
+      this.props.palettes.every(
+        ({ paletteName }) => paletteName.toLowerCase() !== value.toLowerCase()
+      )
+    );
+  }
+
+  handleChange(e) {
+    this.setState({
+      [e.target.name]: e.target.value,
+    });
+  }
+
+  render() {
+    const { open, handleSubmit, handleDrawerOpen } = this.props;
+    const { newPaletteName } = this.state;
+    return (
+      <div>
+        <CssBaseline />
+        <AppBar position='fixed' open={open} color='default'>
+          <Toolbar>
+            <IconButton
+              color='inherit'
+              aria-label='open drawer'
+              onClick={handleDrawerOpen}
+              edge='start'
+              sx={{ mr: 2, ...(open && { display: "none" }) }}>
+              <MenuIcon />
+            </IconButton>
+            <Typography variant='h6' noWrap component='div'>
+              Persistent drawer
+            </Typography>
+            <ValidatorForm onSubmit={() => handleSubmit(newPaletteName)}>
+              <TextValidator
+                label='Palette Name'
+                name='newPaletteName'
+                value={newPaletteName}
+                onChange={this.handleChange}
+                validators={["required", "isPaletteNameUnique"]}
+                errorMessages={["Enter Palette Name", "Name already used"]}
+              />
+              <Button variant='contained' color='primary' type='submit'>
+                Save Palette
+              </Button>
+            </ValidatorForm>
+            <Link to='/'>
+              <Button variant='contained' color='secondary'>
+                Go Back
+              </Button>
+            </Link>
+          </Toolbar>
+        </AppBar>
+      </div>
+    );
+  }
+}
+
+export default PaletteFormNav;
